# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Trails — Magical Musical Submarine API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /onboarding/next-pair:
    get:
      operationId: getNextPair
      tags: [onboarding]
      summary: Get the next discriminative pair for pairwise ranking
      parameters:
        - name: userId
          in: query
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Next pair or null if done
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PairResponse"
  /pair:
    post:
      operationId: submitPair
      tags: [onboarding]
      summary: Submit pairwise ranking result
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PairInput"
      responses:
        "200":
          description: Pair recorded
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OkResponse"
...
  /dive:
    post:
      operationId: createDive
      tags: [dives]
      summary: Start a new dive
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/DiveInput"
      responses:
        "201":
          description: Dive created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Dive"
    get:
      operationId: listDives
      tags: [dives]
      summary: List all dives for a user
      parameters:
        - name: userId
          in: query
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: List of dives
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Dive"
  /dive/detail:
    get:
      operationId: loadDive
      tags: [dives]
      summary: Get a dive with all steps
      parameters:
        - name: diveId
          in: query
          required: true
          schema:
            type: integer
        - name: userId
          in: query
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Dive details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DiveDetail"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /recap:
    get:
      operationId: loadRecap
      tags: [dives]
      summary: Get previously-on recap for a dive
      parameters:
        - name: diveId
          in: query
          required: true
          schema:
            type: integer
        - name: userId
          in: query
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Recap narrative
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RecapResponse"
  /directions:
    post:
      operationId: getDirections
      tags: [dives]
      summary: Generate LLM hypothesis and 3 themed directions
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/DirectionsInput"
      responses:
        "200":
          description: Directions generated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DirectionsResponse"
  /step:
    post:
      operationId: chooseStep
      tags: [dives]
      summary: Record chosen direction and create a dive step
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/StepInput"
      responses:
        "201":
          description: Step created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DiveStep"
  /recommend:
    post:
      operationId: getRecommendations
      tags: [dives]
      summary: Run Propose→Resolve→Narrate for a dive step
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RecommendInput"
      responses:
        "200":
          description: Recommendations with narratives
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Recommendation"
...
components:
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required: [error]
    OkResponse:
      type: object
      properties:
        ok:
          type: boolean
      required: [ok]

    # --- Users ---
...
    PairResponse:
      type: object
      properties:
        done:
          type: boolean
        aMbid:
          type: ["string", "null"]
        aTitle:
          type: ["string", "null"]
        aArtist:
          type: ["string", "null"]
        bMbid:
          type: ["string", "null"]
        bTitle:
          type: ["string", "null"]
        bArtist:
          type: ["string", "null"]
        pairIndex:
          type: ["integer", "null"]
        totalPairs:
          type: ["integer", "null"]
      required: [done, aMbid, aTitle, aArtist, bMbid, bTitle, bArtist, pairIndex, totalPairs]
    PairInput:
      type: object
      properties:
        userId:
          type: integer
        aMbid:
          type: string
        bMbid:
          type: string
        result:
          type: integer
          minimum: -2
          maximum: 2
          description: "-2=strongly A, -1=A, 0=tie, 1=B, 2=strongly B"
      required: [userId, aMbid, bMbid, result]

    # --- Portrait ---
...
    Dive:
      type: object
      properties:
        id:
          type: integer
        userId:
          type: integer
        name:
          type: string
        status:
          type: string
          enum: [active, archived]
        stepCount:
          type: integer
        createdAt:
          type: string
          format: date-time
      required: [id, userId, name, status, stepCount, createdAt]
    DiveInput:
      type: object
      properties:
        userId:
          type: integer
        name:
          type: string
      required: [userId, name]

    # Stored directions as a flat serializable shape (avoids zod.looseObject)
    StoredDirections:
      type: object
      properties:
        hypothesis:
          type: string
        directions:
          type: array
          items:
            $ref: "#/components/schemas/Direction"
        wellTroddenDirection:
          $ref: "#/components/schemas/Direction"
      required: [hypothesis, directions, wellTroddenDirection]

    # Stored links as flat serializable shape
    StoredLinks:
      type: object
      properties:
        spotify:
          type: ["string", "null"]
        youtube:
          type: ["string", "null"]
        appleMusic:
          type: ["string", "null"]
        source:
          type: string
          enum: [odesli, search_fallback]
      required: [spotify, youtube, appleMusic, source]
    DiveStep:
      type: object
      properties:
        id:
          type: integer
        diveId:
          type: integer
        seq:
          type: integer
        hypothesisText:
          type: ["string", "null"]
        directionsJson:
          oneOf:
            - $ref: "#/components/schemas/StoredDirections"
            - type: "null"
        chosenDirection:
          type: ["string", "null"]
        recommendations:
          type: array
          items:
            $ref: "#/components/schemas/Recommendation"
        pathRating:
          oneOf:
            - $ref: "#/components/schemas/PathRating"
            - type: "null"
        createdAt:
          type: string
          format: date-time
      required: [id, diveId, seq, hypothesisText, directionsJson, chosenDirection, recommendations, pathRating, createdAt]
    DiveDetail:
      type: object
      properties:
        id:
          type: integer
        userId:
          type: integer
        name:
          type: string
        status:
          type: string
          enum: [active, archived]
        steps:
          type: array
          items:
            $ref: "#/components/schemas/DiveStep"
        createdAt:
          type: string
          format: date-time
      required: [id, userId, name, status, steps, createdAt]
    Direction:
      type: object
      properties:
        label:
          type: string
        rationale:
          type: string
        isWellTrodden:
          type: boolean
      required: [label, rationale, isWellTrodden]
    DirectionsInput:
      type: object
      properties:
        userId:
          type: integer
        diveId:
          type: integer
      required: [userId, diveId]
    DirectionsResponse:
      type: object
      properties:
        hypothesis:
          type: string
        directions:
          type: array
          items:
            $ref: "#/components/schemas/Direction"
        wellTroddenDirection:
          $ref: "#/components/schemas/Direction"
      required: [hypothesis, directions, wellTroddenDirection]
    StepInput:
      type: object
      properties:
        userId:
          type: integer
        diveId:
          type: integer
        chosenDirection:
          type: string
        hypothesisText:
          type: string
        directionsJson:
          $ref: "#/components/schemas/StoredDirections"
      required: [userId, diveId, chosenDirection, hypothesisText, directionsJson]
    RecapResponse:
      type: object
      properties:
        recap:
          type: string
        stepCount:
          type: integer
        avgScore:
          type: ["number", "null"]
      required: [recap, stepCount, avgScore]

    # --- Recommendations ---
    Recommendation:
      type: object
      properties:
        id:
          type: integer
        diveStepId:
          type: integer
        type:
          type: string
          enum: [track, album]
        mbid:
          type: string
        title:
          type: string
        artist:
          type: string
        year:
          type: ["integer", "null"]
        narrativeText:
          type: ["string", "null"]
        linksJson:
          oneOf:
            - $ref: "#/components/schemas/StoredLinks"
            - type: "null"
        artworkUrl:
          type: ["string", "null"]
        arm:
          type: string
          enum: [llm, well_trodden]
        likelyKnown:
          type: ["string", "null"]
        latestRating:
          oneOf:
            - $ref: "#/components/schemas/Rating"
            - type: "null"
        createdAt:
          type: string
          format: date-time
      required: [id, diveStepId, type, mbid, title, artist, year, narrativeText, linksJson, artworkUrl, arm, likelyKnown, latestRating, createdAt]
    RecommendInput:
      type: object
      properties:
        stepId:
          type: integer
        userId:
          type: integer
      required: [stepId, userId]

    # --- Streaming links ---
...
    Rating:
      type: object
      properties:
        id:
          type: integer
        recId:
          type: integer
        listenState:
          type: string
          enum: [listened, skipped, known]
        score:
          type: ["number", "null"]
        ratedAt:
          type: string
          format: date-time
      required: [id, recId, listenState, score, ratedAt]
...
    PathRating:
      type: object
      properties:
        id:
          type: integer
        diveStepId:
          type: integer
        score:
          type: number
        ratedAt:
          type: string
          format: date-time
      required: [id, diveStepId, score, ratedAt]
...